refactor: http driver provider to use withOptions and withConfig functions
