+++ packages/ngworker/lumberjack/http-driver/src/lib/configuration/lumberjack-http-driver.provider.ts
@@ -1,0 +1,11 @@
+import { Provider } from '@angular/core';
+
+import { lumberjackLogDriverToken } from '@ngworker/lumberjack';
+
+import { LumberjackHttpDriver } from '../log-drivers/lumberjack-http.driver';
+
+export const lumberjackHttpDriverProvider: Provider = {
+  provide: lumberjackLogDriverToken,
+  useClass: LumberjackHttpDriver,
+  multi: true,
+};
